refactor: cleanup group schema
