+++ schemas/group.py
@@ -13,12 +13,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============================================================================
-from pydantic import BaseModel
+from pydantic import BaseModel, field_validator
 
 from schemas import ORMBaseModel
+from services.validation.geospatial import validate_wkt_geometry
 
 
 # -------- CREATE ----------
+class CreateGroup(BaseModel):
+    """
+    Schema for creating a group.
+    """
+
+    name: str
+    description: str | None = None
+    parent_group_id: int | None = None
+    project_area: str | None = None
+
+    @classmethod
+    @field_validator("project_area")
+    def validate_area_is_wkt(cls, wkt):
+        return validate_wkt_geometry(wkt)
+
+
 # -------- RESPONSE --------
 class GroupResponse(ORMBaseModel):
     """
@@ -42,6 +59,7 @@ class UpdateGroup(BaseModel):
     name: str | None = None
     description: str | None = None
     parent_group_id: int | None = None
+    project_area: str | None = None
 
 
 # ============= EOF =============================================
