+++ schemas/location.py
@@ -19,6 +19,7 @@ from pydantic import BaseModel, field_validator
 from shapely import wkt
 
 from schemas import ORMBaseModel
+from services.validation.geospatial import validate_wkt_geometry
 
 """
 TODO
@@ -26,23 +27,6 @@ TODO
 Create common validator classes to be shared amongst create and update schemas.
 Since many fields are optional in the update schemas set check_fields=False in the field_validator.
 """
-
-
-def validate_wkt_geometry(value: str | None) -> str | None:
-    """
-    Validate that the provided string is a valid WKT geometry.
-    Raises ValueError if the geometry is invalid.
-    """
-    if value is None:
-        return value
-
-    try:
-        geometry = wkt.loads(value)
-        if not geometry.is_valid:
-            raise ValueError("WKT geometry is not topologically valid")
-        return value
-    except Exception as e:
-        raise ValueError(f"Invalid WKT geometry: {e}")
 
 
 # -------- CREATE ----------
@@ -59,22 +43,6 @@ class CreateLocation(BaseModel):
     @classmethod
     @field_validator("point")
     def validate_point_is_wkt(cls, wkt):
-        return validate_wkt_geometry(wkt)
-
-
-class CreateGroup(BaseModel):
-    """
-    Schema for creating a group.
-    """
-
-    name: str
-    description: str | None = None
-    parent_group_id: int | None = None
-    project_area: str | None = None
-
-    @classmethod
-    @field_validator("project_area")
-    def validate_area_is_wkt(cls, wkt):
         return validate_wkt_geometry(wkt)
